修正負傷沒有正確讀取

+++ GameObjects/GameObjects/GameScenario.cs
@@ -3276,7 +3276,7 @@
                 }
                 try
                 {
-                    person.InjureRate = (int)reader["InjureRate"];
+                    person.InjureRate = (float)reader["InjureRate"];
                 }
                 catch 
                 {
